OMI_Voice_Portal_/requirements.txt

flask
google-cloud-texttospeech
requests
gunicorn
